+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Adds a confirmation screen for first-time bidders - yuki24
 * Adds a new color `red100` (`#F7625A`) to the theme - yuki24
 * Adds a `<CheckBox>` component - yuki24
+* Adds support for viewing conditions of sale - ash
 
 ### 1.4.6